Extract Klotski micro-framework from main game logic
Klotski is named after a Polish sliding-tile puzzle. It is a
micro-framework sitting on top of monogame, supporting the
easy creation of tile-based 2D platform games.

Ricochet-specific logic should go in the Ricochet project,
whereas game physics and general mechanics should go in Klotski.
The Ricochet project should not need to reference monogame.